backend/requirements.txt
bcrypt
flask
pandas
cors
flask-cors
uuid
pymongo
python-dotenv
datetime
pytz
bson
openpyxl
flask-jwt-extended